Improve PeerManager handling of duplicate connections
This commits improves the handling the case where both
peers are simultaneously trying to connect.

The peers need to both choose to close the same connection.
To handle this, the peers will compare their identities.
The peer with the greater id will close the incoming connection
while the peer with the lesser id will close the outbound
connection.

The commit also fixes outbound reconnection and improves
logging to include the peer_id.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>